refactor: separate experiments names from rest of configuration
This creates a problem where we now try to synchronize same name
across the two places. To address this, we have dataclasses that
enumerate expected names. This way, if a new experiment is added
to the config file, it will raise an error that names file was
not updated and vice versa.