WIP - testing HabitatSim observations
In an effort to get to the bottom of the differences between various OSs
and environments when running the tests, I'm adding a test to try to
verify that the observations that the HabitatSim is getting match across
test environments.

At the moment, I'm just planning to manually test these, comparing my
local Linux environment with the GitHub Actions environment.