refactor: MontyExperiment context manager part 2
This is the second phase of [the work I proposed on Discourse][forum] to
make MontyExperiment into a context manager.

This moves the config setup into a constructor for the class, leaving
the part that initializes the model and datasets in the
`setup_experiment` method. Then we call the simplified setup method in
the context manager `__enter__` method.

The rest of this is updating the call sites to use the new context
manager style more completely.

[forum]: https://thousandbrains.discourse.group/t/investigation-into-linux-test-failures-and-possible-improvements/468