chore: add tests for ProfileExperimentMixin
While working on #220, we realized that this mixin might be affected by
the changes there. Upon investigation, I noticed that it was lacking any
tests to make sure at a bare minimum that it works, so I'm adding some here.

I also added an `__init_subclass__` method to it to check and ensure the
required invariants are true when it is used.