style: enable ruff format checking
We were under the impression that `ruff check` checked for formatting, but that is under a separate subcommand `ruff format`. This change formats the files that have drifted from the formatting we want, and updates the GH checks to make sure we keep it that way going forward.

After this is merged, we'll add the merge commit hash to the `.git-blame-ignore-revs` file to keep the commit history clear in GH and other tools that support that feature.