chore: check for pending review requests
Whenever someone requests a re-review, we aren't checking for that with the check-assignee-approval check, so this updates the check to do that.

I also realize a lot of the information we're using here is available in the context, so we don't need to make a separate API call to fetch the pull request object.

The new behaviour is:
- check for empty assignees
- check for non-maintainer assignees
- check for pending assignee reviews
- check latest reviews were approvals