chore: initial uv environment setup
As part of our work to add a second simulator, we're setting up the
project so its environment can be managed using `uv`. This is in
parallel to the environment managed by Conda.

For now Conda will be used when using Habitat Sim, and the `uv` project
will be used when working with the new simulator (most likely to be
MuJoCo).

This also involved making the tests that depend on Habitat Sim get
skipped when that dependency is not installed.