docs: update config to include constructors
We're currently not including "dunder" methods, e.g. `__init__` in our
generated API documentation. This changes the Sphinx configuration to
include a specific list of them, starting with `__init__`.

We don't want to set this to include all of them, because it will
include attributes like `__dict__` and `__weakref__` which we don't
want. Plus sphinx-build hangs when trying to document `__module__` in
one of the HabitatSim modules.