ci: fix broken uv builds
This involved a bit of work, but we need to specify the dependencies
that Conda seems to include with PyTorch by default that the PyPI
version doesn't.

To get it to work, we have to add an extra "find links" for the Pyg
repository to pull in the prebuilt versions. Unfortunately, those
prebuilt versions (at least for the versions we're pinning to) only
exists for MacOS x86_64. Similar to the Conda environment, we need to
force `uv` to use an x86_64 build of Python to get this to work. When
creating the environment, we need to specify the exact version of Python
we want.

```bash
uv venv --python cpython-3.9.22-macos-x86_64-none
```