fix: correct data type in MontyRequestHandler
Problem: if a query parameter is added to a PUT request, it breaks the
data type detection. A depth request that includes query parameters
would be treated as RGB, which is the default when the path is not
recognized as depth-typed.

Solution: Now the url is parsed using urlparse, to identify whether the
data is depth or RGB, by separating the queries from the path.

More specifics: Note that in HTTPRequestHandler a self.path value is
available. It is a complete URL, not a file path, including query params:
https://docs.python.org/3/library/http.server.html#http.server.BaseHTTPRequestHandler.path

That's why this uses urlparse instead, for example:
>>> path='foo://example.com:8042/over/there?name=ferret'
>>> import os
>>> os.path.basename(path)
'there?name=ferret'
>>> from urllib.parse import urlparse
>>> urlparse(path)
ParseResult(scheme='foo', netloc='example.com:8042', path='/over/there', params='', query='name=ferret', fragment='')
>>> parsed = urlparse(path)
>>> os.path.basename(parsed.path)
'there'